Adding schematic and BOM template
Adding APSS-A3.SchDot schematic template which is an initial idea containing the APSS logo and relevant project parameters to be filled out. Adding APSS BOM Template.xlsx which is an initial idea for the APSS BOM template (this will need further work to link back to the schematic file project parameters in case we go forward with it).