Redesign of Programming Adapter.
Added: 
1x 4-Pin Picoclasp header
1x 8-Pin Picoclasp header

Changed: 
Size of the board has increased to fit new components and traces.
Complete layout of the board.
Position of the JTAG header relative to the pogopins (Moved further away)

Question: Are the two 4 pin headers at the bottom of the board too far away from the edge? Or are they fine where they currently are?